[#IOPID-1920] Remove duplicated code in test suite
