+++ businesslogic/__tests__/analytics-profiles.test.ts
@@ -2,23 +2,19 @@ import * as TE from "fp-ts/TaskEither";
 import { NonNegativeInteger } from "@pagopa/ts-commons/lib/numbers";
 import { EmailString, NonEmptyString } from "@pagopa/ts-commons/lib/strings";
 
-import { pipe } from "fp-ts/lib/function";
-import * as RA from "fp-ts/ReadonlyArray";
-
 import { getAnalyticsProcessorForDocuments } from "../analytics-publish-documents";
 
 import { aFiscalCode } from "../../__mocks__/services.mock";
-import { sha256 } from "../../utils/pdv";
 import { RetrievedProfile } from "@pagopa/io-functions-commons/dist/src/models/profile";
 import { PreferredLanguageEnum } from "@pagopa/io-functions-commons/dist/generated/definitions/PreferredLanguage";
 import { ReminderStatusEnum } from "@pagopa/io-functions-commons/dist/generated/definitions/ReminderStatus";
 import { ServicesPreferencesModeEnum } from "@pagopa/io-functions-commons/dist/generated/definitions/ServicesPreferencesMode";
 import { PushNotificationsContentTypeEnum } from "@pagopa/io-functions-commons/dist/generated/definitions/PushNotificationsContentType";
 import {
+  aCosmosMetadata,
+  anError,
   mockGetPdvId,
-  mockSendMessageViaQueue,
   mockSendMessageViaTopic,
-  mockTrackException,
   trackerAdapter
 } from "../__mocks__/processor.mock";
 import {
@@ -26,18 +22,13 @@ import {
   getMainAdapter,
   getFallbackAdapterWithFilter
 } from "../__mocks__/processor.mock";
+import {
+  getEnricherFailureExpecter,
+  getKafkaProducerFailureExpects,
+  getSuccessValidListExpects
+} from "../__mocks__/test-case";
 
 // Data
-
-const aTopic = "a-topic";
-const anError = new Error("An error");
-
-const aCosmosMetadata = {
-  _etag: "_etag",
-  _rid: "_rid",
-  _self: "xyz",
-  _ts: 1
-};
 
 const aRetrievedProfile: RetrievedProfile = {
   ...aCosmosMetadata,
@@ -90,20 +81,7 @@ describe("publish", () => {
     // When
     await processorAdapter.process(documents)();
     // Then
-    expect(mockGetPdvId).toHaveBeenCalledTimes(2);
-    expect(mockSendMessageViaTopic).toHaveBeenCalledTimes(1);
-    expect(mockSendMessageViaTopic).toHaveBeenCalledWith({
-      messages: documents.map(document => ({
-        value: JSON.stringify({
-          ...document,
-          // enriched values
-          userPDVId: sha256(document.fiscalCode)
-        })
-      })),
-      topic: aTopic
-    });
-    expect(mockSendMessageViaQueue).toHaveBeenCalledTimes(0);
-    expect(mockTrackException).toHaveBeenCalledTimes(0);
+    getSuccessValidListExpects(documents);
   });
 
   it("GIVEN a valid list of profile and a not working Kafka Producer Client, WHEN processing the list, THEN send it to the queue", async () => {
@@ -122,25 +100,7 @@ describe("publish", () => {
     // When
     await processorAdapter.process(documents)();
     // Then
-    expect(mockGetPdvId).toHaveBeenCalledTimes(2);
-    expect(mockSendMessageViaTopic).toHaveBeenCalledTimes(1);
-    expect(mockSendMessageViaQueue).toHaveBeenCalledTimes(2);
-    pipe(
-      documents,
-      RA.mapWithIndex((i, document) =>
-        expect(mockSendMessageViaQueue).toHaveBeenNthCalledWith(
-          i + 1,
-          Buffer.from(
-            JSON.stringify({
-              ...document
-              // DO NOT store pdvId values
-              // userPDVId: sha256(document.fiscalCode)
-            })
-          ).toString("base64")
-        )
-      )
-    );
-    expect(mockTrackException).toHaveBeenCalledTimes(0);
+    getKafkaProducerFailureExpects(documents);
   });
 
   it("GIVEN a valid list of profiles and a transient error on content enricher, WHEN processing the list, THEN send it to the queue", async () => {
@@ -157,24 +117,6 @@ describe("publish", () => {
     // When
     await processorAdapter.process(documents)();
     // Then
-    expect(mockGetPdvId).toHaveBeenCalledTimes(2);
-    expect(mockSendMessageViaTopic).toHaveBeenCalledTimes(1);
-    expect(mockSendMessageViaQueue).toHaveBeenCalledTimes(1);
-    pipe(
-      [documents[0]],
-      RA.mapWithIndex((i, document) =>
-        expect(mockSendMessageViaQueue).toHaveBeenNthCalledWith(
-          i + 1,
-          Buffer.from(
-            JSON.stringify({
-              ...document
-              // DO NOT store pdvId values
-              // userPDVId: sha256(document.fiscalCode)
-            })
-          ).toString("base64")
-        )
-      )
-    );
-    expect(mockTrackException).toHaveBeenCalledTimes(0);
+    getEnricherFailureExpecter(documents);
   });
 });
